extensions for response added

+++ backend/src/PetFamily.Application/Volunteers/Delete/DeleteVolunteerService.cs
@@ -1,5 +1,8 @@
 ﻿using CSharpFunctionalExtensions;
+using FluentValidation;
 using Microsoft.Extensions.Logging;
+using PetFamily.Application.Extensions;
+using PetFamily.Application.Volunteers.Create;
 using PetFamily.Domain.Shared.Errors;
 using System;
 using System.Collections.Generic;
@@ -9,24 +12,35 @@ using System.Threading.Tasks;
 
 namespace PetFamily.Application.Volunteers.Delete
 {
-    public class DeleteVolunteerService : IDeleteVolunteerService
+    public class DeleteVolunteerService 
     {
         private readonly IVolunteerRepository _volunteerRepository;
         private readonly ILogger<DeleteVolunteerService> _logger;
+        private readonly IValidator<DeleteVolunteerCommand> _validator;
 
-        public DeleteVolunteerService(IVolunteerRepository repository, ILogger<DeleteVolunteerService> logger)
+
+        public DeleteVolunteerService(
+            IVolunteerRepository repository, 
+            ILogger<DeleteVolunteerService> logger, 
+            IValidator<DeleteVolunteerCommand> validator)
         {
             _volunteerRepository = repository;
             _logger = logger;
-
+            _validator = validator;
         }
 
-        public async Task<Result<Guid, Error>> Delete(DeleteVolunteerRequest request, CancellationToken cancellationToken)
+        public async Task<Result<Guid, ErrorList>> Delete(DeleteVolunteerCommand command, CancellationToken cancellationToken)
         {
-            var volunteerResult = await _volunteerRepository.GetById(request.Id, cancellationToken);
+            var validationResult = await _validator.ValidateAsync(command, cancellationToken);
+            if(validationResult.IsValid ==false)
+            {
+                return validationResult.ToErrorList();
+            }
+
+            var volunteerResult = await _volunteerRepository.GetById(command.Id, cancellationToken);
             if (volunteerResult.IsFailure)
             {
-                return volunteerResult.Error;
+                return new ErrorList([volunteerResult.Error]);
             }
 
             var result =  _volunteerRepository.Delete(volunteerResult.Value, cancellationToken);
